requirements.txt
PyPDF2==3.0.1
python-markdown==3.4.4
pytesseract==0.3.10
Pillow==10.0.0
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.42
json5==0.9.14
python-dotenv==1.0.0
pdfplumber==0.9.0
python-docx==0.8.11
python-pptx==0.6.21
unstructured==0.10.30
unstructured[pdf]==0.10.30
unstructured[docx]==0.10.30
unstructured[md]==0.10.30
chromadb==0.4.24
openai==1.14.3